+++ src/examples/train.py
@@ -6,7 +6,6 @@ Launch this with torchrun:
     torchrun --nproc-per-node=4 src/examples/train.py run_name [OVERRIDES...]
 """
 
-import json
 import sys
 from dataclasses import dataclass
 from typing import List, cast
@@ -14,7 +13,7 @@ from typing import List, cast
 from olmo_core.config import Config, DType
 from olmo_core.data import MemMapDatasetConfig, TokenizerConfig
 from olmo_core.distributed.parallel import DataParallelConfig, DataParallelType
-from olmo_core.distributed.utils import get_rank, init_hybrid_shard_mesh
+from olmo_core.distributed.utils import init_hybrid_shard_mesh
 from olmo_core.nn.transformer import TransformerConfig
 from olmo_core.optim import AdamWConfig, CosWithWarmup
 from olmo_core.train import (
@@ -24,6 +23,7 @@ from olmo_core.train import (
 )
 from olmo_core.train.callbacks import (
     CheckpointerCallback,
+    ConfigSaverCallback,
     GPUMemoryMonitorCallback,
     GradClipperCallback,
     SchedulerCallback,
@@ -99,6 +99,7 @@ def build_config(run_name: str, overrides: List[str]) -> ExperimentConfig:
                 enabled=False,  # change to true to enable
             ),
         )
+        .with_callback("config_saver", ConfigSaverCallback())
     )
 
     return ExperimentConfig(
@@ -122,11 +123,10 @@ def main(run_name: str, overrides: List[str]):
     dataset = config.dataset.build()
     trainer = config.trainer.build(model, optim, dataset)
 
-    # Save config to W&B and file.
-    if get_rank() == 0:
-        config_dict = config.as_config_dict()
-        cast(WandBCallback, trainer.callbacks["wandb"]).config = config_dict
-        trainer.write_file("config.json", json.dumps(config_dict, indent=2))
+    # Save config to W&B and each checkpoint dir.
+    config_dict = config.as_config_dict()
+    cast(WandBCallback, trainer.callbacks["wandb"]).config = config_dict
+    cast(ConfigSaverCallback, trainer.callbacks["config_saver"]).config = config_dict
 
     # Train.
     trainer.fit()